# docs/swagger.yaml
basePath: /
definitions:
...
  ds.Flights:
    properties:
      date_Flight:
        type: string
      id_Application:
        type: integer
      id_Cosmodrom_Begin:
        type: integer
      id_Ship:
        type: integer
      id_cosmodrom_End:
        type: integer
    type: object
...
  ds.delete_flight:
    properties:
      idApp:
        type: integer
      idShip:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: Starship's flights
  license:
    name: AS IS (NO WARRANTY)
  title: SpaceY
  version: "1.0"
paths:
  /api/application:
    delete:
      consumes:
      - application/json
      description: Удалить космолет из заявки
      parameters:
      - description: Данные для удаления полета из заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.delete_flight'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно удалено
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Удалить космолет из заявки
      tags:
      - Полет
...
  /api/flights/cosmodrom/begin:
    put:
      consumes:
      - application/json
      description: Установить космодром вылета
      parameters:
      - description: Данные для удаления полета из заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.Flights'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Установить космодром вылета
      tags:
      - Полет
  /api/flights/cosmodrom/end:
    put:
      consumes:
      - application/json
      description: Установить космодром прилета
      produces:
      - application/json
      responses:
        "200":
          description: Успешно
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Установить космодром прилета
      tags:
      - Полет
  /api/flights/cosmodroms:
    get:
      consumes:
      - application/json
      description: Получить список космодромов
      produces:
      - application/json
      responses:
        "200":
          description: Успешно
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить список космодромов
      tags:
      - Полет
  /api/flights/date:
    put:
      consumes:
      - application/json
      description: Установить дату полета
      parameters:
      - description: Данные для удаления полета из заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.Flights'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Установить дату полета
      tags:
      - Полет
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"